I added powerups to increase the players health and shield capacity. Now, when the player begins the game, the player's ship has 10 units of health and shields. I added two power-up to increase both the ship's health and shield capacity from 10 to 20 and from 20 to 40.
When the power-ups appear is dependent on the player's score. After a certain score both power-ups to increase the health and shield capacity from 10 to 20 will begin to appear and the same goes for the power-ups to increase the ship's capacity from 20 to 40.

I also made some other power-ups appear when certain conditions are met. For instance only full health and full shield power-ups will appear if the player's health or shields are less than full. The firepower power-ups will not appear if the player has the maximum fire power.