The biggest change in this commit is refactoring the powerup code. I removed the methods responsible for spawning powerups and added them to a new PowerupManager. Also, creating a powerup manager allowed me to clean up a few utility methods from the PlayerEntity class.
I also finished removing the remaining EntityImage variables and replacing their references with the ImageManager.

I also tweaked how the sounds are called. I removed the calls for the asteroid breaking apart from the method breakAsteroid and moved them to the collision methods. I also changed when an asteroid sound is played. When an asteroid hits the players shield, only the sound of the shield getting hit is played. Otherwise the asteroid sound is played. This makes the game a bit mroe cleaner by not trying to play multiple sounds at the same time, which has resulted in some of the sounds not being played.