First, I tested the refactoring of the UFO code from the previous commit. I made sure the respective UFOs were being launched. During testing I noticed there was a problem with the UFOs position on the screen. As it turns out when I changed the constructor signature for the UFOEntity I removed the image parameter. When I was setting the image in one of the UFO sub classes after calling the super constructor, the posiiton was getting reset to 0,0 (via the EntityImage setImage method). I changed back the constructor for UFOEntity to include the image parameter and everything is not working correctly.
I also improved the logic for determining the UFO position, direction and headign vector in the class UFOEntity. The logic seems a bit more straight forward and the heading vector will be between  set of specified angles to make the game play better.

I also cleaned up some of the code to display the HUD by using the StaticText class for some of the messages displayed to the screen. This reduces some of the code in the main game class Asteroids.